pyproject.toml
[tool.poetry]
name = "hierarchical_data"
version = "0.1.0"
description = ""
authors = ["Sean Bailey <sepabailey@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
